# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin MVC Template.",
        "title": "Gin MVC Template",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mritd/ginmvc",
            "email": "mritd@linux.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": ":8080",
    "paths": {
        "/api/v1/user/login": {
            "post": {
                "description": "user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "user login",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"success\",\"timestamp\":1570889247}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"user not registered or password is wrong\",\"timestamp\":1570889272}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    },
                    "500": {
                        "description": "{\"message\":\"invalid connection\",\"timestamp\":1570887977}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "description": "user logout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "user logout",
                "responses": {
                    "200": {
                        "description": "{\"message\":\"success\",\"timestamp\":1570889247}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    },
                    "500": {
                        "description": "{\"message\":\"error message\",\"timestamp\":1570887977}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "description": "user register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "user register",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"success\",\"timestamp\":1570887849}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"user email [mritd@linux.com] already register\",\"timestamp\":1570887884}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    },
                    "500": {
                        "description": "{\"message\":\"invalid connection\",\"timestamp\":1570887977}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "health check",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor API"
                ],
                "summary": "health check",
                "responses": {
                    "200": {
                        "description": "{\"message\":\"success\",\"timestamp\":1570889247}",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CommonResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "models.NullBool": {
            "type": "object"
        },
        "models.NullInt32": {
            "type": "object"
        },
        "models.NullInt64": {
            "type": "object"
        },
        "models.NullString": {
            "type": "object"
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullInt64"
                },
                "email": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullString"
                },
                "id": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullInt32"
                },
                "lock": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullBool"
                },
                "loginTime": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullInt64"
                },
                "mobile": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullString"
                },
                "name": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullString"
                },
                "password": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullString"
                },
                "salt": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullString"
                },
                "updateTime": {
                    "type": "object",
                    "$ref": "#/definitions/models.NullInt64"
                }
            }
        }
    }
}